requirements.txt
flask
psycopg2-binary
python-docx
gunicorn
python-dotenv
